Move entities' convertation, representation and comparison to entity module
