+++ test/selenium/src/lib/constants/messages.py
@@ -1,10 +1,9 @@
 # Copyright (C) 2017 Google Inc.
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 """Constants and procedures to make formatted messages."""
-
 import copy
 
-from lib.entities import entity
+from lib.entities.entity import Entity
 
 
 class CommonMessages(object):
@@ -27,7 +26,7 @@ class AssertionMessages(CommonMessages):
   err_three_entities_diff = (
       _line + "\nFULL:" + _triple_diff + "\nSAME:" + _triple_diff + "\nDIFF:" +
       _triple_diff)
-  all_entities_classes = tuple(entity.Entity().all_entities_classes())
+  _all_entities_classes = tuple(Entity.all_entities_classes())
 
   @classmethod
   def diff_error_msg(cls, left, right):
@@ -41,14 +40,15 @@ class AssertionMessages(CommonMessages):
     """Check if entities are instances of entities and have needed attributes
     and keys to make result error comparison message.
     """
-    return (isinstance(left and right, cls.all_entities_classes) and
-            hasattr(left and right, "diff_info") and
-            getattr(left, "diff_info") and getattr(right, "diff_info"))
+    return (
+        isinstance(left and right, cls._all_entities_classes) and
+        hasattr(left and right, "diff_info") and
+        getattr(left, "diff_info") and getattr(right, "diff_info"))
 
   @classmethod
   def set_entities_diff_info(cls, left, right):
     """Get and set entities diff info attributes info."""
-    comparison = entity.Representation.compare_entities(left, right)
+    comparison = Entity.compare_entities(left, right)
     left.diff_info = comparison["self_diff"]
     right.diff_info = comparison["other_diff"]
 
@@ -60,14 +60,14 @@ class AssertionMessages(CommonMessages):
     # comparison of entities
     assertion_error_msg = cls.err_common.format(left, right)
     # processing of entity
-    if isinstance(left and right, cls.all_entities_classes):
+    if isinstance(left and right, cls._all_entities_classes):
       if not cls.is_entities_have_err_info(left, right):
         cls.set_entities_diff_info(left, right)
       assertion_error_msg = cls.diff_error_msg(left, right)
     # processing list of entities
     if (isinstance(left and right, list) and
-            all(isinstance(_left and _right, cls.all_entities_classes)
-                for _left, _right in zip(left, right))):
+        all(isinstance(_left and _right, cls._all_entities_classes)
+            for _left, _right in zip(left, right))):
       assertion_error_msg = ""
       for _left, _right in zip(sorted(left), sorted(right)):
         if not cls.is_entities_have_err_info(_left, _right):
@@ -84,14 +84,14 @@ class AssertionMessages(CommonMessages):
     # comparison of entities
     assertion_error_msg = cls.err_contains.format(left, right)
     # processing of entity
-    if isinstance(left and right, cls.all_entities_classes):
+    if isinstance(left and right, cls._all_entities_classes):
       if not cls.is_entities_have_err_info(left, right):
         cls.set_entities_diff_info(left, right)
       assertion_error_msg = cls.diff_error_msg(left, right)
     # processing list of entities
-    if (isinstance(left, cls.all_entities_classes) and
+    if (isinstance(left, cls._all_entities_classes) and
             isinstance(right, list) and
-            all(isinstance(_right, cls.all_entities_classes)
+            all(isinstance(_right, cls._all_entities_classes)
                 for _right in right)):
       assertion_error_msg = ""
       for _right in right:
