+++ test/selenium/src/lib/entities/entity.py
@@ -4,27 +4,21 @@
 # pylint: disable=too-many-arguments
 # pylint: disable=too-few-public-methods
 
+import copy
 from datetime import datetime
+
 import pytest
 
-from lib.utils import string_utils
+from lib.utils import string_utils, help_utils
 
 
 class Representation(object):
   """Class that contains methods to update Entity."""
   # pylint: disable=import-error
-
   diff_info = None  # {"equal": {"atr7": val7, ...}, "diff": {"atr3": val3}}
   attrs_names_to_compare = None
   attrs_names_to_repr = None
   core_attrs_names_to_repr = ["type", "title", "id", "href", "url", "slug"]
-
-  def extended_equal(self, other):
-    """Extended equal procedure fore self and other entities."""
-    comparison = Representation.compare_entities(self, other)
-    self.diff_info = comparison["self_diff"]
-    other.diff_info = comparison["other_diff"]
-    return comparison["is_equal"]
 
   @classmethod
   def get_attrs_names_for_entities(cls, entity=None):
@@ -36,6 +30,13 @@ class Representation(object):
     all_entities_attrs_names = string_utils.convert_list_elements_to_list(
         [entity_cls().__dict__.keys() for entity_cls in all_entities_cls])
     return list(set(all_entities_attrs_names))
+
+  def __repr__(self):
+    # exclude attributes witch are used for REST interaction forming
+    # pylint: disable=not-an-iterable
+    return str(dict(zip(self.attrs_names_to_repr,
+                        [getattr(self, attr_name_to_repr) for attr_name_to_repr
+                         in self.attrs_names_to_repr])))
 
   @staticmethod
   def items_of_remap_keys():
@@ -70,20 +71,182 @@ class Representation(object):
     result_remap_items.update(csv_remap_items)
     return string_utils.dict_keys_to_upper_case(result_remap_items)
 
+  @staticmethod
+  def convert_objs_repr_to_dict(obj_or_objs):
+    """Convert objects' representation to dictionary 'obj.attr_name' =
+    'attr_value' to dictionary or list of dictionaries with items
+    {'attr_name': 'attr_value'}.
+    """
+    if obj_or_objs:
+      if isinstance(obj_or_objs, list):
+        if (all(not isinstance(_, dict) and
+                not isinstance(_, (str, unicode, int)) and
+                _ for _ in obj_or_objs)):
+          obj_or_objs = [_.__dict__ for _ in obj_or_objs]
+      else:
+        if (not isinstance(obj_or_objs, dict) and
+                not isinstance(obj_or_objs, (str, unicode, int))):
+          obj_or_objs = obj_or_objs.__dict__
+      return obj_or_objs
+
+  @staticmethod
+  def convert_dict_to_obj_repr(dic):
+    """Convert dictionary to Entity representation (dictionary's keys and
+    values to attributes names and attributes values).
+    """
+    # pylint: disable=expression-not-assigned
+    # pylint: disable=consider-iterating-dictionary
+    entity = Entity()
+    [delattr(entity, k) for k in entity.__dict__.keys()]
+    [setattr(entity, k, v) for k, v in dic.iteritems()]
+    return entity
+
   def repr_ui(self):
     """Convert entity's attributes values from REST like to UI like
     representation.
     """
-    from lib.entities import entities_factory
-    return (entities_factory.EntitiesFactory().
-            convert_objs_repr_from_rest_to_ui(obj_or_objs=self))
+    return self.convert_objs_repr_from_rest_to_ui(obj_or_objs=self)
+
+  @classmethod  # noqa: ignore=C901
+  def convert_objs_repr_from_rest_to_ui(cls, obj_or_objs):
+    """Convert object's or objects' attributes values from REST like
+    (dict or list of dict) representation to UI like with unicode.
+    Examples:
+    None to None, u'Ex' to u'Ex', [u'Ex1', u'Ex2', ...] to u'Ex1, Ex2',
+    {'name': u'Ex', ...} to u'Ex',
+    [{'name': u'Ex1', ...}, {'name': u'Ex2', ...}] to u'Ex1, Ex2'
+    """
+    # pylint: disable=too-many-locals
+    # pylint: disable=undefined-loop-variable
+    # pylint: disable=invalid-name
+    def convert_obj_repr_from_rest_to_ui(obj):
+      """Convert object's attributes from REST to UI like representation."""
+      def convert_attr_value_from_dict_to_unicode(attr_name, attr_value):
+        """Convert attribute value from dictionary to unicode representation
+        (get value by key from dictionary 'attr_value' where key determine
+        according to 'attr_name').
+        """
+        if isinstance(attr_value, dict):
+          converted_attr_value = attr_value
+          if attr_name in [
+              "contact", "manager", "owners", "assessor", "creator",
+              "verifier", "created_by", "modified_by", "Assessor", "Creator",
+              "Verifier"
+          ]:
+            converted_attr_value = unicode(attr_value.get("name"))
+          if attr_name in ["custom_attribute_definitions", "program", "audit",
+                           "objects_under_assessment"]:
+            converted_attr_value = (
+                unicode(attr_value.get("title")) if
+                attr_name != "custom_attribute_definitions" else
+                {attr_value.get("id"): attr_value.get("title").upper()}
+            )
+          if attr_name in ["custom_attribute_values"]:
+            converted_attr_value = {attr_value.get("custom_attribute_id"):
+                                    attr_value.get("attribute_value")}
+          if obj_attr_name == "comments":
+            converted_attr_value = {
+                k: (string_utils.convert_str_to_datetime(v) if
+                    k == "created_at" and isinstance(v, unicode) else v)
+                for k, v in attr_value.iteritems()
+                if k in ["modified_by", "created_at", "description"]}
+          return converted_attr_value
+      origin_obj = copy.deepcopy(obj)
+      for obj_attr_name in obj.__dict__.keys():
+        # 'Ex', u'Ex', 1, None to 'Ex', u'Ex', 1, None
+        obj_attr_value = (obj.assignees.get(obj_attr_name.title()) if (
+            obj_attr_name in ["assessor", "creator", "verifier"] and
+            "assignees" in obj.__dict__.keys())
+            else getattr(obj, obj_attr_name))
+        # u'2017-06-07T16:50:16' and u'2017-06-07 16:50:16' to datetime
+        if (obj_attr_name in ["updated_at", "created_at"] and
+                isinstance(obj_attr_value, unicode)):
+          obj_attr_value = string_utils.convert_str_to_datetime(obj_attr_value)
+        if isinstance(obj_attr_value, dict) and obj_attr_value:
+          # to "assignees" = {"Assessor": [], "Creator": [], "Verifier": []}
+          if obj_attr_name == "assignees":
+            obj_attr_value = {
+                k: ([convert_attr_value_from_dict_to_unicode(k, _v)
+                     for _v in v] if isinstance(v, list) else
+                    convert_attr_value_from_dict_to_unicode(k, v))
+                for k, v in obj_attr_value.iteritems()
+                if k in ["Assessor", "Creator", "Verifier"]}
+          # {'name': u'Ex1', 'type': u'Ex2', ...} to u'Ex1'
+          else:
+            obj_attr_value = convert_attr_value_from_dict_to_unicode(
+                obj_attr_name, obj_attr_value)
+        # [el1, el2, ...] or [{item1}, {item2}, ...] to [u'Ex1, u'Ex2', ...]
+        if (isinstance(obj_attr_value, list) and
+                all(isinstance(item, dict) for item in obj_attr_value)):
+          obj_attr_value = [
+              convert_attr_value_from_dict_to_unicode(obj_attr_name, item) for
+              item in obj_attr_value]
+        setattr(obj, obj_attr_name, obj_attr_value)
+      # merge "custom_attribute_definitions" and "custom_attribute_values"
+      obj_cas_attrs_names = [
+          "custom_attributes", "custom_attribute_definitions",
+          "custom_attribute_values"]
+      if set(obj_cas_attrs_names).issubset(obj.__dict__.keys()):
+        cas_def = obj.custom_attribute_definitions
+        cas_val = obj.custom_attribute_values
+        # form CAs values of CAs definitions exist but CAs values not, or CAs
+        # definitions have different then CAs values lengths
+        if (cas_def and
+                (not cas_val or (isinstance(cas_def and cas_val, list)) and
+                 len(cas_def) != len(cas_val))):
+          from lib.entities.entities_factory import (
+              CustomAttributeDefinitionsFactory)
+          cas_val_dicts_keys = ([_.keys()[0] for _ in cas_val] if
+                                isinstance(cas_val, list) else [None])
+          _cas_val = [
+              {k: v} for k, v in
+              CustomAttributeDefinitionsFactory.generate_ca_values(
+                  list_ca_def_objs=origin_obj.custom_attribute_definitions,
+                  is_none_values=True).iteritems()
+              if k not in cas_val_dicts_keys]
+          cas_val = _cas_val if not cas_val else cas_val + _cas_val
+        cas_def_dict = (
+            dict([_def.iteritems().next() for _def in cas_def]) if
+            (isinstance(cas_def, list) and
+             all(isinstance(_def, dict)
+                 for _def in cas_def)) else {None: None})
+        cas_val_dict = (
+            dict([_val.iteritems().next() for _val in cas_val]) if
+            (isinstance(cas_def, list) and
+             all(isinstance(_def, dict)
+                 for _def in cas_def)) else {None: None})
+        cas = string_utils.merge_dicts_by_same_key(cas_def_dict, cas_val_dict)
+        setattr(obj, "custom_attributes", cas)
+      return obj
+    return help_utils.execute_method_according_to_plurality(
+        obj_or_objs=obj_or_objs, types=Entity.all_entities_classes(),
+        method_name=convert_obj_repr_from_rest_to_ui)
 
   def repr_snapshot(self, parent_obj):
     """Convert entity's attributes values to Snapshot representation."""
-    from lib.entities import entities_factory
-    return (entities_factory.EntitiesFactory().
-            convert_objs_repr_to_snapshot(
-                obj_or_objs=self, parent_obj=parent_obj))
+    return (self.convert_objs_repr_to_snapshot(
+        obj_or_objs=self, parent_obj=parent_obj))
+
+  @classmethod  # noqa: ignore=C901
+  def convert_objs_repr_to_snapshot(cls, obj_or_objs, parent_obj):
+    """Convert object's or objects' attributes values to Snapshot
+    representation.
+    Retrieved values will be used for: 'id'.
+    Set values will be used for: 'title, 'type', 'slug', 'href'.
+    """
+    def convert_obj_repr_to_snapshot(origin_obj, parent_obj):
+      """Convert object's attributes to Snapshot representation."""
+      from lib.service import rest_service
+      origin_obj = copy.deepcopy(origin_obj)
+      snapshoted_obj = (
+          rest_service.ObjectsInfoService().get_snapshoted_obj(
+              origin_obj=origin_obj, paren_obj=parent_obj))
+      origin_obj.__dict__.update(
+          {k: v for k, v in snapshoted_obj.__dict__.iteritems()})
+      return origin_obj
+    return help_utils.execute_method_according_to_plurality(
+        obj_or_objs=obj_or_objs, types=Entity.all_entities_classes(),
+        method_name=convert_obj_repr_to_snapshot, parent_obj=parent_obj)
 
   def update_attrs(self, is_replace_attrs=True, is_allow_none=True,
                    is_replace_dicts_values=False, **attrs):
@@ -92,47 +255,95 @@ class Representation(object):
     If 'is_replace_values_of_dicts' then update values of dicts in list which
     is value of particular object's attribute name.
     """
-    from lib.entities import entities_factory
-    return (entities_factory.EntitiesFactory().
-            update_objs_attrs_values_by_entered_data(
-                obj_or_objs=self, is_replace_attrs_values=is_replace_attrs,
-                is_allow_none_values=is_allow_none,
-                is_replace_values_of_dicts=is_replace_dicts_values, **attrs))
+    return (self.update_objs_attrs_values_by_entered_data(
+        obj_or_objs=self, is_replace_attrs_values=is_replace_attrs,
+        is_allow_none_values=is_allow_none,
+        is_replace_values_of_dicts=is_replace_dicts_values, **attrs))
 
-  def compare_entities(self, other):
-    """Extended compare of entities: 'self_entity' and 'other_entity' according
-    to specific 'attrs_names_to_compare' and return:
-    - 'is_equal' - True if entities equal else False;
-    - 'self_diff' - 'equal' and 'diff' parts of 'self_entity' after compare;
-    - 'other_diff' - 'equal' and 'diff' parts of 'other_entity' after compare.
+  @classmethod
+  def update_objs_attrs_values_by_entered_data(
+      cls, obj_or_objs, is_replace_attrs_values=True,
+      is_allow_none_values=True, is_replace_values_of_dicts=False, **arguments
+  ):
+    """Update object or list of objects ('obj_or_objs') attributes values by
+    manually entered data if attribute name exist in 'attrs_names' witch equal
+    to 'all_objs_attrs_names' according to dictionary of attributes and values
+    '**arguments'. If 'is_replace_attrs_values' then replace attributes values,
+    if not 'is_replace_attrs_values' then update (merge) attributes values
+    witch should be lists. If 'is_allow_none_values' then allow to set None
+    object's attributes values, and vice versa.
+    If 'is_replace_values_of_dicts' then update values of dicts in list which
+    is value of particular object's attribute name:
+    (**arguments is attr={'key1': 'new_value2', 'key2': 'new_value2'}).
     """
-    # pylint: disable=not-an-iterable
-    is_equal = False
-    self_equal, self_diff, other_equal, other_diff = {}, {}, {}, {}
-    if (isinstance(self, other.__class__) and
-            self.attrs_names_to_compare == other.attrs_names_to_compare):
-      for attr_name in self.attrs_names_to_compare:
-        self_attr_value = None
-        other_attr_value = None
-        if (attr_name in self.__dict__.keys() and
-                attr_name in other.__dict__.keys()):
-          self_attr_value = getattr(self, attr_name)
-          other_attr_value = getattr(other, attr_name)
-          is_equal = self.is_attrs_equal(
-              attr_name, self_attr_value, other_attr_value)
-        if is_equal:
-          self_equal[attr_name] = self_attr_value
-          other_equal[attr_name] = other_attr_value
-        else:
-          self_diff[attr_name] = {"val": self_attr_value,
-                                  "type": type(self_attr_value)}
-          other_diff[attr_name] = {"val": other_attr_value,
-                                   "type": type(other_attr_value)}
-      is_equal = self_diff == other_diff == {}
-    return {"is_equal": is_equal,
-            "self_diff": {"equal": self_equal, "diff": self_diff},
-            "other_diff": {"equal": other_equal, "diff": other_diff}
-            }
+    # pylint: disable=expression-not-assigned
+    # pylint: disable=invalid-name
+    def update_obj_attrs_values(obj, is_replace_attrs_values,
+                                is_allow_none_values, **arguments):
+      """Update object's attributes values."""
+      for obj_attr_name in arguments:
+        if (obj_attr_name in
+                Entity.get_attrs_names_for_entities(obj.__class__)):
+          _obj_attr_value = arguments.get(obj_attr_name)
+          condition = (True if is_allow_none_values else _obj_attr_value)
+          if condition and not is_replace_values_of_dicts:
+            # convert repr from objects to dicts exclude datetime objects
+            obj_attr_value = (
+                cls.convert_objs_repr_to_dict(_obj_attr_value) if
+                not isinstance(_obj_attr_value, datetime) else _obj_attr_value)
+            if not is_replace_attrs_values:
+              origin_obj_attr_value = getattr(obj, obj_attr_name)
+              obj_attr_value = (
+                  dict(origin_obj_attr_value.items() + obj_attr_value.items())
+                  if obj_attr_name == "custom_attributes" else
+                  string_utils.convert_to_list(origin_obj_attr_value) +
+                  string_utils.convert_to_list(obj_attr_value))
+            setattr(obj, obj_attr_name, obj_attr_value)
+          if is_replace_values_of_dicts and isinstance(_obj_attr_value, dict):
+            obj_attr_value = string_utils.exchange_dicts_items(
+                transform_dict=_obj_attr_value,
+                dicts=string_utils.convert_to_list(
+                    getattr(obj, obj_attr_name)),
+                is_keys_not_values=False)
+            obj_attr_value = (
+                obj_attr_value if isinstance(getattr(obj, obj_attr_name), list)
+                else obj_attr_value[0])
+            setattr(obj, obj_attr_name, obj_attr_value)
+      return obj
+    return help_utils.execute_method_according_to_plurality(
+        obj_or_objs=obj_or_objs, types=Entity.all_entities_classes(),
+        method_name=update_obj_attrs_values,
+        is_replace_attrs_values=is_replace_attrs_values,
+        is_allow_none_values=is_allow_none_values, **arguments)
+
+  @classmethod
+  def filter_objs_attrs(cls, obj_or_objs, attrs_to_include):
+    """Make objects's copy and filter objects's attributes (delete attributes
+    from objects witch not in list'attrs_to_include').
+    'objs' can be list of objects or object.
+    """
+    # pylint: disable=expression-not-assigned
+    def filter_obj_attrs(obj, attrs_to_include):
+      """Filter one object's attributes."""
+      obj = copy.deepcopy(obj)
+      [delattr(obj, obj_attr) for obj_attr in obj.__dict__.keys()
+       if obj_attr not in attrs_to_include]
+      return obj
+    return ([filter_obj_attrs(obj, attrs_to_include) for obj in obj_or_objs] if
+            isinstance(obj_or_objs, list) else
+            filter_obj_attrs(obj_or_objs, attrs_to_include))
+
+  def __eq__(self, other):
+    """Extended equal procedure fore self and other entities."""
+    comparison = Representation.compare_entities(self, other)
+    self.diff_info = comparison["self_diff"]
+    other.diff_info = comparison["other_diff"]
+    return comparison["is_equal"]
+
+  @staticmethod
+  def attrs_values_types_error(self_attr, other_attr, expected_types):
+    raise ValueError("'{}' have to be isinstance of classes: {}\n".
+                     format((self_attr, other_attr), expected_types))
 
   @classmethod
   def is_attrs_equal(cls, attr_name, self_attr_value, other_attr_value):
@@ -202,10 +413,40 @@ class Representation(object):
           self_attr=self_comments, other_attr=other_comments,
           expected_types=(list.__name__, type(None).__name__))
 
-  @staticmethod
-  def attrs_values_types_error(self_attr, other_attr, expected_types):
-    raise ValueError("'{}' have to be isinstance of classes: {}\n".
-                     format((self_attr, other_attr), expected_types))
+  def compare_entities(self, other):
+    """Extended compare of entities: 'self_entity' and 'other_entity' according
+    to specific 'attrs_names_to_compare' and return:
+    - 'is_equal' - True if entities equal else False;
+    - 'self_diff' - 'equal' and 'diff' parts of 'self_entity' after compare;
+    - 'other_diff' - 'equal' and 'diff' parts of 'other_entity' after compare.
+    """
+    # pylint: disable=not-an-iterable
+    is_equal = False
+    self_equal, self_diff, other_equal, other_diff = {}, {}, {}, {}
+    if (isinstance(self, other.__class__) and
+            self.attrs_names_to_compare == other.attrs_names_to_compare):
+      for attr_name in self.attrs_names_to_compare:
+        self_attr_value = None
+        other_attr_value = None
+        if (attr_name in self.__dict__.keys() and
+                attr_name in other.__dict__.keys()):
+          self_attr_value = getattr(self, attr_name)
+          other_attr_value = getattr(other, attr_name)
+          is_equal = self.is_attrs_equal(
+              attr_name, self_attr_value, other_attr_value)
+        if is_equal:
+          self_equal[attr_name] = self_attr_value
+          other_equal[attr_name] = other_attr_value
+        else:
+          self_diff[attr_name] = {"val": self_attr_value,
+                                  "type": type(self_attr_value)}
+          other_diff[attr_name] = {"val": other_attr_value,
+                                   "type": type(other_attr_value)}
+      is_equal = self_diff == other_diff == {}
+    return {"is_equal": is_equal,
+            "self_diff": {"equal": self_equal, "diff": self_diff},
+            "other_diff": {"equal": other_equal, "diff": other_diff}
+            }
 
   @classmethod
   def issue_assert(cls, expected_objs, actual_objs, issue_msg,
@@ -239,16 +480,6 @@ class Representation(object):
         for expected_exclude_attr, actual_exclude_attr
         in zip(expected_exclude_attrs, actual_exclude_attrs))
         else pytest.xfail(reason=issue_msg))
-
-  def __eq__(self, other):
-    return self.extended_equal(other)
-
-  def __repr__(self):
-    # exclude attributes witch are used for REST interaction forming
-    # pylint: disable=not-an-iterable
-    return str(dict(zip(self.attrs_names_to_repr,
-                        [getattr(self, attr_name_to_repr) for attr_name_to_repr
-                         in self.attrs_names_to_repr])))
 
 
 class Entity(Representation):
@@ -345,7 +576,6 @@ class CustomAttributeEntity(Representation):
   # pylint: disable=redefined-builtin
   # pylint: disable=too-many-instance-attributes
   __hash__ = None
-  entity = Entity()
   attrs_names_to_compare = [
       "type", "title", "definition_type", "attribute_type", "mandatory"]
   attrs_names_to_repr = [
