+++ test/selenium/src/lib/service/webui_service.py
@@ -7,7 +7,6 @@ import re
 
 from lib import factory
 from lib.constants import objects, url, path, messages, roles, element, regex
-from lib.entities.entities_factory import EntitiesFactory
 from lib.entities.entity import Entity
 from lib.page import dashboard
 from lib.page.widget.info_widget import SnapshotableInfoPanel
@@ -79,7 +78,7 @@ class BaseWebUiService(object):
                   "*" in val):
             scope[key] = val.replace("*", "")
     return [
-        EntitiesFactory.update_objs_attrs_values_by_entered_data(
+        Entity.update_objs_attrs_values_by_entered_data(
             obj_or_objs=factory_obj, is_allow_none_values=False, **scope) for
         scope, factory_obj in zip(list_scopes_to_convert, list_factory_objs)]
